Refactor project structure and add PromptLayerBase
